backend/pyproject.toml
[project]
name = "jobsearch-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "sqlalchemy>=2.0",
    "alembic>=1.14",
    "psycopg2-binary>=2.9",
    "playwright>=1.49",
    "beautifulsoup4>=4.12",
    "httpx>=0.28",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "html5lib>=1.1",
]

[project.optional-dependencies]
llm = [
    "langchain-openai>=0.3",
    "langchain-core>=0.3",
]

[tool.setuptools.packages.find]
where = ["."]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
